Reduce memory requirements.
Add pyro.compute_3d_ voronoi method: reduce memory requirements, allow points exclusion from the calculation, and specify the properties to be retrieved for each cell.

The calculation is performed in a slightly different way so that the properties of each cell are retrieved right after its identification and not at the end of the tessellation.